JBR-339 refactor money
