+++ src/app/money/category-picker/money-cat-picker.component.ts
@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Category, ICategory} from '../money-category';
 import {MoneyService} from '../money.service';
-import {JbAccount} from '../money-account';
+import {JbAccount} from '../money-jbaccount';
 
 export class MoneyCategoryPickerSelectableOption {
     category: ICategory;
