+++ src/app/money/money-file.ts
@@ -1,3 +1,3 @@
 export interface IFile {
-    file: string;
+    filename: string;
 }
