+++ src/app/money/money-list.component.html
@@ -240,7 +240,7 @@
                 <table class="table trn-table trn-table-files" *ngIf="files && files.length">
                     <tr *ngFor="let nextFile of files">
                         <td class="col-8 trn-table-col-file">
-                            {{nextFile.file}}
+                            {{nextFile.filename}}
                         </td>
                         <td class="col-1 trn-table-col-opts">
                             <div class="row">
